requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
click==6.7
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-extensions==3.1.0
django-heroku==0.3.1
djangorestframework==3.12.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.7
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==0.23
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plaid-python==5.0.0
prompt-toolkit==3.0.9
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.7.3
python-dotenv==0.14.0
pytz==2020.5
requests==2.20.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.24.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
